ESC-318 fix compiler warnings and move service tests into correct package - `services` not `service`
